explicitly install kexec-tools on all platforms
on trusty, kexec-tools is no longer a dependency of linux-crashdump. We
now install it explicitly to make sure we have the `kdump` service, so
we can actually *load* the crashkernel.

n.b.: we cannot test for service enabled because beaker-puppet checks
for runlevel 3, and kdump starts in 2.

This commit fixes #2